Move helper funding reducer
